// ui/src/app/service/persistence/dto/jwt-registry.dto.ts
// Shared DTO: Copy in back-end and front-end should be identical

export interface TokenCreateDto {
  token: string;
}

export interface JwtRegistryClaimsDto {
  client_id: string;
  exp: number;
  jti: string;
  sub: string;
}

export interface JwtRegistryDto {
  id: string;
  accountId: string;
  blocked?: boolean;
  claims: JwtRegistryClaimsDto;
  createdUserId: string;
  createdAt: string;
}